parameter 타입 변경

+++ ImageDrawView/src/main/java/com/github/crazyatom/subscaleimagedrawview/listener/ShowSnackbarListener.java
@@ -5,6 +5,6 @@ package com.github.crazyatom.subscaleimagedrawview.listener;
  */
 
 public interface ShowSnackbarListener {
-    void onShown(final float height);
+    void onShown(final int height);
     void onDismissed();
 }
